requirements.txt
absl-py==1.4.0
addict==2.4.0
ansi2html==1.8.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.0
cycler==0.11.0
dash==2.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dearpygui==1.9.1
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
Flask==2.2.5
fonttools==4.42.1
fsspec==2023.10.0
fvcore==0.1.5.post20221221
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
h5py==3.9.0
huggingface-hub==0.18.0
icecream==2.1.3
idna==3.4
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.10
ipython==8.12.2
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
lazy_loader==0.3
lpips==0.1.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nbformat==5.7.0
nerfacc==0.3.3
nest-asyncio==1.5.7
networkx==3.1
ninja==1.11.1
numpy==1.23.1
oauthlib==3.2.2
open3d==0.17.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.16.1
plyfile==1.0.1
portalocker==2.7.0
prettytable==3.8.0
prompt-toolkit==3.0.39
protobuf==4.24.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
Pygments==2.16.1
pyhocon==0.3.60
PyMCubes==0.1.4
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.8
torch==1.13.1+cu116
torch-tb-profiler==0.4.1
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.66.1
traitlets==5.9.0
transforms3d==0.4.1
trimesh==3.22.5
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
yacs==0.1.8
zipp==3.16.2